refactor: split syslog/syslog.go: extract writeloop + drain (#540)
syslog/syslog.go was 929 lines — over the project's ~300-line
guideline. Split the writeLoop goroutine and its helpers (batch
state, flush logic, writeEntry, no-retry shutdown drain) into a
new sibling file. Pure refactor; no behaviour change, no public
API change.

Kept in syslog.go (651 lines):
- Package doc, imports, severity mapping, ReconnectRecorder
  interface, syslogEntry struct, Output struct
- Public API: New, Write, WriteWithMetadata, enqueue, Close,
  ReportsDelivery, Name, DestinationKey, LastDeliveryNanos
- connect, boundedTLSDialer (reconnect machinery)
- recordSuccess (called from three sites)
- handleWriteFailure, retryAfterReconnect (reconnect-on-error path)

Moved to syslog_writeloop.go (307 lines):
- writeLoop (top-level loop)
- batchState struct + methods (newBatchState, append,
  thresholdReached)
- flushAndReset, handleShutdownDrain, flushBatch, resetSyslogTimer
- writeEntry
- errSyslogNotConnected sentinel
- drainBatchNoRetry, drainRemainingNoRetry, drainOne (shutdown
  no-retry path)

Plan deviates from issue's proposed three-file split (syslog.go +
syslog_writeloop.go + syslog_reconnect.go): putting reconnect in a
third file would split tightly-coupled code (writeEntry calls
handleWriteFailure; handleWriteFailure calls connect; recordSuccess
is called from three sites across boundaries). The two-file split
delivers the structural benefit without the cross-file coupling
cost.

Both files remain over the 300-line target — accepted per the same
rationale as audit.go (#540): further fragmentation harms cohesion.

Verification:
- go build ./... → clean
- go test -race -count=1 ./syslog → green (12s)
- go test -race -count=1 ./... → green (full suite)

Refs #540.
